+++ integration_tests/test_gravity.py
@@ -1,31 +1,25 @@
 import json
 
 import pytest
-import sha3
-import toml
 from eth_account.account import Account
-from eth_utils import abi, to_checksum_address
+from eth_utils import abi
 from hexbytes import HexBytes
-from pystarport import ports
 from web3.exceptions import BadFunctionCallOutput
 
-from .gorc import GoRc
-from .network import GravityBridge, setup_cronos, setup_geth
+from .gravity_utils import prepare_gravity, setup_cosmos_erc20_contract
+from .network import setup_cronos, setup_geth
 from .utils import (
     ACCOUNTS,
     ADDRS,
     CONTRACTS,
     KEYS,
-    add_ini_sections,
     approve_proposal,
     deploy_contract,
-    deploy_erc20,
-    dump_toml,
     eth_to_bech32,
-    get_contract,
     multiple_send_to_cosmos,
     send_to_cosmos,
     send_transaction,
+    setup_token_mapping,
     w3_wait_for_new_blocks,
     wait_for_fn,
     wait_for_new_blocks,
@@ -39,41 +33,6 @@ Account.enable_unaudited_hdwallet_features()
 def cronos_crc21_abi():
     path = CONTRACTS["ModuleCRC21"]
     return json.load(path.open())["abi"]
-
-
-def gorc_config(keystore, gravity_contract, eth_rpc, cosmos_grpc, metrics_listen):
-    return {
-        "keystore": str(keystore),
-        "gravity": {
-            "contract": gravity_contract,
-            "fees_denom": "basetcro",
-        },
-        "ethereum": {
-            "key_derivation_path": "m/44'/60'/0'/0/0",
-            "rpc": eth_rpc,
-        },
-        "cosmos": {
-            "gas_price": {
-                "amount": 5000000000000,
-                "denom": "basetcro",
-            },
-            "gas_limit": 500000,
-            "grpc": cosmos_grpc,
-            "key_derivation_path": "m/44'/60'/0'/0/0",
-            "prefix": "crc",
-            "msg_batch_size": 10,
-        },
-        "metrics": {
-            "listen_addr": metrics_listen,
-        },
-    }
-
-
-def update_gravity_contract(tomlfile, contract):
-    with open(tomlfile) as fp:
-        obj = toml.load(fp)
-    obj["gravity"]["contract"] = contract
-    tomlfile.write_text(dump_toml(obj))
 
 
 def check_auto_deployment(cli, denom, cronos_w3, recipient, amount):
@@ -110,132 +69,18 @@ def get_id_from_receipt(receipt):
 
 
 @pytest.fixture(scope="module")
-def geth(tmp_path_factory):
-    "start-geth"
-    for network in setup_geth(tmp_path_factory.mktemp("geth"), 8555):
-        yield network.w3
+def custom_geth(tmp_path_factory):
+    yield from setup_geth(tmp_path_factory.mktemp("geth"), 8555)
 
 
 @pytest.fixture(scope="module", params=[True, False])
-def cronos(request, tmp_path_factory):
-    """start-cronos
-    params: enable_auto_deployment
-    """
+def custom_cronos(request, tmp_path_factory):
     yield from setup_cronos(tmp_path_factory.mktemp("cronos"), 26600, request.param)
 
 
 @pytest.fixture(scope="module")
-def gravity(cronos, geth):
-    """
-    - set-delegator-keys
-    - deploy gravity contract
-    - start orchestrator
-    """
-    chain_id = "cronos_777-1"
-
-    # set-delegate-keys
-    for i, val in enumerate(cronos.config["validators"]):
-        # generate gorc config file
-        gorc_config_path = cronos.base_dir / f"node{i}/gorc.toml"
-        grpc_port = ports.grpc_port(val["base_port"])
-        metrics_port = 3000 + i
-        gorc_config_path.write_text(
-            dump_toml(
-                gorc_config(
-                    cronos.base_dir / f"node{i}/orchestrator_keystore",
-                    "",  # to be filled later after the gravity contract deployed
-                    geth.provider.endpoint_uri,
-                    f"http://localhost:{grpc_port}",
-                    f"127.0.0.1:{metrics_port}",
-                )
-            )
-        )
-
-        gorc = GoRc(gorc_config_path)
-
-        # generate new accounts on both chain
-        gorc.add_eth_key("eth")
-        gorc.add_eth_key("cronos")  # cronos and eth key derivation are the same
-
-        # fund the orchestrator accounts
-        eth_addr = to_checksum_address(gorc.show_eth_addr("eth"))
-        print("fund 0.1 eth to address", eth_addr)
-        send_transaction(geth, {"to": eth_addr, "value": 10**17}, KEYS["validator"])
-        acc_addr = gorc.show_cosmos_addr("cronos")
-        print("fund 100cro to address", acc_addr)
-        rsp = cronos.cosmos_cli().transfer(
-            "community", acc_addr, "%dbasetcro" % (100 * (10**18))
-        )
-        assert rsp["code"] == 0, rsp["raw_log"]
-
-        cli = cronos.cosmos_cli(i)
-        val_addr = cli.address("validator", bech="val")
-        val_acct_addr = cli.address("validator")
-        nonce = int(cli.account(val_acct_addr)["base_account"]["sequence"])
-        signature = gorc.sign_validator("eth", val_addr, nonce)
-        rsp = cli.set_delegate_keys(
-            val_addr, acc_addr, eth_addr, HexBytes(signature).hex(), from_=val_acct_addr
-        )
-        assert rsp["code"] == 0, rsp["raw_log"]
-    # wait for gravity signer tx get generated
-    wait_for_new_blocks(cli, 2)
-
-    # create admin account and fund it
-    admin, _ = Account.create_with_mnemonic()
-    print("fund 0.1 eth to address", admin.address)
-    send_transaction(geth, {"to": admin.address, "value": 10**17}, KEYS["validator"])
-
-    # deploy gravity contract to geth
-    gravity_id = cli.query_gravity_params()["params"]["gravity_id"]
-    signer_set = cli.query_latest_signer_set_tx()["signer_set"]["signers"]
-    powers = [int(signer["power"]) for signer in signer_set]
-    threshold = int(2**32 * 0.66)  # gravity normalize the power to [0, 2**32]
-    eth_addresses = [signer["ethereum_address"] for signer in signer_set]
-    assert sum(powers) >= threshold, "not enough validator on board"
-
-    contract = deploy_contract(
-        geth,
-        CONTRACTS["Gravity"],
-        (gravity_id.encode(), threshold, eth_addresses, powers, admin.address),
-    )
-    print("gravity contract deployed", contract.address)
-
-    # make all the orchestrator "Relayer" roles
-    k_relayer = sha3.keccak_256()
-    k_relayer.update(b"RELAYER")
-    for _, address in enumerate(eth_addresses):
-        set_role_tx = contract.functions.grantRole(
-            k_relayer.hexdigest(), address
-        ).build_transaction({"from": admin.address})
-        set_role_receipt = send_transaction(geth, set_role_tx, admin.key)
-        print("set_role_tx", set_role_receipt)
-
-    # start orchestrator:
-    # a) add process into the supervisord config file
-    # b) reload supervisord
-    programs = {}
-    for i, val in enumerate(cronos.config["validators"]):
-        # update gravity contract in gorc config
-        gorc_config_path = cronos.base_dir / f"node{i}/gorc.toml"
-        update_gravity_contract(gorc_config_path, contract.address)
-
-        programs[f"program:{chain_id}-orchestrator{i}"] = {
-            "command": (
-                f'gorc -c "{gorc_config_path}" orchestrator start '
-                "--cosmos-key cronos --ethereum-key eth --mode AlwaysRelay"
-            ),
-            "environment": "RUST_BACKTRACE=full",
-            "autostart": "true",
-            "autorestart": "true",
-            "startsecs": "3",
-            "redirect_stderr": "true",
-            "stdout_logfile": f"%(here)s/orchestrator{i}.log",
-        }
-
-    add_ini_sections(cronos.base_dir / "tasks.ini", programs)
-    cronos.supervisorctl("update")
-
-    yield GravityBridge(cronos, geth, contract)
+def gravity(custom_cronos, custom_geth):
+    yield from prepare_gravity(custom_cronos, custom_geth)
 
 
 def test_gravity_transfer(gravity):
@@ -595,48 +440,15 @@ def test_gravity_cancel_transfer(gravity):
 
 def test_gravity_source_tokens(gravity):
     if not gravity.cronos.enable_auto_deployment:
-        # deploy crc21 contract
+        # deploy contracts
         w3 = gravity.cronos.w3
-        contract = deploy_contract(w3, CONTRACTS["TestERC21Source"])
-
-        # setup the contract mapping
-        cronos_cli = gravity.cronos.cosmos_cli()
-
-        print("crc21 contract", contract.address)
-        denom = f"cronos{contract.address}"
-
-        print("check the contract mapping not exists yet")
-        with pytest.raises(AssertionError):
-            cronos_cli.query_contract_by_denom(denom)
-
-        rsp = cronos_cli.update_token_mapping(
-            denom, contract.address, "DOG", 6, from_="validator"
+        symbol = "DOG"
+        contract, denom = setup_token_mapping(gravity.cronos, "TestERC21Source", symbol)
+        cosmos_erc20_contract = setup_cosmos_erc20_contract(
+            gravity,
+            denom,
+            symbol,
         )
-        assert rsp["code"] == 0, rsp["raw_log"]
-        wait_for_new_blocks(cronos_cli, 1)
-
-        print("check the contract mapping exists now")
-        rsp = cronos_cli.query_denom_by_contract(contract.address)
-        assert rsp["denom"] == denom
-
-        # Create cosmos erc20 contract
-        print("Deploy cosmos erc20 contract on ethereum")
-        tx_receipt = deploy_erc20(
-            gravity.contract, gravity.geth, denom, denom, "DOG", 6, KEYS["validator"]
-        )
-        assert tx_receipt.status == 1, "should success"
-
-        # Wait enough for orchestrator to relay the event
-        wait_for_new_blocks(cronos_cli, 30)
-
-        # Check mapping is done on gravity side
-        cosmos_erc20 = cronos_cli.query_gravity_contract_by_denom(denom)
-        print("cosmos_erc20:", cosmos_erc20)
-        assert cosmos_erc20 != ""
-        cosmos_erc20_contract = get_contract(
-            gravity.geth, cosmos_erc20["erc20"], CONTRACTS["TestERC21Source"]
-        )
-
         # Send token to ethereum
         amount = 1000
         ethereum_receiver = ADDRS["validator"]
